LayoutTests/accessibility/mac/id-class-attributes-expected.txt

content
This tests that the DOM id and the DOM class list can be retrieved

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item.isAttributeSupported('AXDOMIdentifier') is true
PASS item.stringAttributeValue('AXDOMIdentifier') is 'item1'
PASS item.isAttributeSupported('AXDOMClassList') is true
PASS item.classList is 'test, test2'
PASS successfullyParsed is true

TEST COMPLETE
